refactor(server): split oversized server file by responsibility
